+++ x-pack/platform/plugins/shared/triggers_actions_ui/public/application/sections/rules_page/rules_page.tsx
@@ -15,7 +15,7 @@ import { getCreateRuleRoute, getEditRuleRoute } from '@kbn/rule-data-utils';
 import { useGetRuleTypesPermissions } from '@kbn/alerts-ui-shared';
 import { RulesPageTemplate } from './rules_page_template';
 import { useKibana } from '../../../common/lib/kibana';
-import { getAlertingSectionBreadcrumb } from '../../lib/breadcrumb';
+import { getAlertingSectionBreadcrumb, getRulesBreadcrumbWithHref } from '../../lib/breadcrumb';
 import { getCurrentDocTitle } from '../../lib/doc_title';
 import { NON_SIEM_CONSUMERS } from '../alerts_search_bar/constants';
 import type { Section } from '../../constants';
@@ -30,7 +30,7 @@ const RulesPage = () => {
   const {
     chrome: { docTitle },
     setBreadcrumbs,
-    application: { navigateToApp },
+    application: { navigateToApp, getUrlForApp, isAppRegistered },
     http,
     notifications: { toasts },
   } = useKibana().services;
@@ -141,10 +141,18 @@ const RulesPage = () => {
 
   useEffect(() => {
     if (setBreadcrumbs) {
-      setBreadcrumbs([getAlertingSectionBreadcrumb(currentSection || 'rules')]);
+      if (currentSection === 'logs') {
+        const rulesBreadcrumbWithAppPath = getRulesBreadcrumbWithHref(
+          isAppRegistered,
+          getUrlForApp
+        );
+        setBreadcrumbs([rulesBreadcrumbWithAppPath, getAlertingSectionBreadcrumb('logs')]);
+      } else {
+        setBreadcrumbs([getAlertingSectionBreadcrumb('rules')]);
+      }
     }
     docTitle.change(getCurrentDocTitle(currentSection || 'rules'));
-  }, [docTitle, setBreadcrumbs, currentSection]);
+  }, [docTitle, setBreadcrumbs, currentSection, getUrlForApp, isAppRegistered]);
 
   return (
     <RulesPageTemplate
